requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
email-validator==2.2.0
sqlalchemy==2.0.48
alembic==1.18.4
pydantic-settings==2.11.0
python-dotenv==1.1.1
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
